refactor: merge setting pages
