+++ src/pages/User.jsx
@@ -1,29 +1,10 @@
-import { Link, useNavigate } from "react-router-dom";
+import { Link } from "react-router-dom";
 import home from "../assets/home.svg";
 import search from "../assets/search.svg";
 import my from "../assets/my.svg";
-import {
-  Drawer,
-  DrawerClose,
-  DrawerContent,
-  DrawerDescription,
-  DrawerFooter,
-  DrawerHeader,
-  DrawerTitle,
-  DrawerTrigger,
-} from "@/components/ui/drawer";
-import { Button } from "@/components/ui/button";
-import { useContext } from "react";
-import IsLoginContext from "@/components/contexts/IsLoginContext";
+import LogoutDrawer from "@/components/setting/LogoutDrawer";
 
 const User = () => {
-  const { logout } = useContext(IsLoginContext);
-  const navigate = useNavigate();
-
-  const handleLogout = () => {
-    logout();
-    navigate("/");
-  };
   return (
     <div className="flex justify-center h-screen min-h-[600px] bg-background">
       <div className="relative flex flex-col items-center p-14 w-10/12 max-w-2xl h-full bg-white rounded-2xl">
@@ -32,7 +13,7 @@ const User = () => {
         </div>
         <div className="border-b w-full ">
           <Link
-            to="/user/user-info"
+            to="/user/info"
             className="flex py-5 w-full font-bold transition-all hover:underline"
           >
             회원정보 수정
@@ -40,7 +21,7 @@ const User = () => {
         </div>
         <div className="border-b w-full ">
           <Link
-            to="/user/category-select"
+            to="/user/category"
             className="flex py-5 w-full font-bold transition-all hover:underline"
           >
             선호 주제 변경
@@ -48,50 +29,18 @@ const User = () => {
         </div>
         <div className="border-b w-full ">
           <Link
-            to="/user/publisher-select"
+            to="/user/publisher"
             className="flex py-5 w-full font-bold transition-all hover:underline"
           >
             뉴스 구독 관리
           </Link>
         </div>
         <div className="border-b w-full ">
-          <Drawer>
-            <DrawerTrigger>
-              <div className="flex py-5 w-full font-bold transition-all hover:underline ">
-                로그아웃
-              </div>
-            </DrawerTrigger>
-            <DrawerContent>
-              <div className="mx-auto w-full max-w-xl">
-                <DrawerHeader className="mt-4 ">
-                  <DrawerTitle className="text-2xl font-bold mb-4">
-                    로그아웃을 하시겠습니까?
-                  </DrawerTitle>
-                  <DrawerDescription>
-                    새로운 소식을 받을 수 없게 됩니다.
-                  </DrawerDescription>
-                </DrawerHeader>
-                <DrawerFooter>
-                  <div className="flex gap-6">
-                    <Button
-                      onClick={handleLogout}
-                      variant="destructive"
-                      className="w-1/2"
-                    >
-                      로그아웃
-                    </Button>
-                    <DrawerClose className="w-1/2">
-                      <Button>취소</Button>
-                    </DrawerClose>
-                  </div>
-                </DrawerFooter>
-              </div>
-            </DrawerContent>
-          </Drawer>
+          <LogoutDrawer />
         </div>
         <div className="border-b w-full ">
           <Link
-            to="/user/account-delete"
+            to="/user/delete"
             className="flex py-5 w-full font-bold transition-all hover:underline"
           >
             회원 탈퇴
